Minor syntax error within Course revision
Corrected a minor syntax error that appeared within one of the final
revisions that added approximately 6 more courses to the Course Table.
The error was minor in that if gone undetected the updated database
would have a total of 6 or so missing courses that would probably only
affect a single major.